+++ src/test/java/de/alpharogroup/meanbean/factories/LocalDateFactoryTest.java
@@ -26,25 +26,27 @@ package de.alpharogroup.meanbean.factories;
 
 import org.junit.jupiter.api.Test;
 
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.MatcherAssert.assertThat;
+import java.time.LocalDate;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
- * The unit test class for the class {@link StringArrayFactory}.
+ * The unit test class for the class {@link LocalDateFactory}.
  */
-public class StringArrayFactoryTest
+public class LocalDateFactoryTest
 {
 
 	/**
-	 * Test for method {@link StringArrayFactory#create()}
+	 * Test for method {@link LocaleFactory#create()}
 	 */
-	@Test public void create()
+	@Test public void testCreate()
 	{
-		String[] expected;
-		String[] actual;
-		StringArrayFactory factory = new StringArrayFactory();
-		expected = new String[] { "foo", "bar", "john", "doe" };
+		LocalDate expected;
+		LocalDate actual;
+		LocalDateFactory factory = new LocalDateFactory();
+		expected = LocalDate.now();
 		actual = factory.create();
-		assertThat(actual, equalTo(expected));
+		assertEquals(actual, expected);
 	}
+
 }