+++ src/test/java/de/alpharogroup/meanbean/factories/LocalDateTimeFactoryTest.java
@@ -26,25 +26,32 @@ package de.alpharogroup.meanbean.factories;
 
 import org.junit.jupiter.api.Test;
 
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.MatcherAssert.assertThat;
+import java.time.LocalDateTime;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
- * The unit test class for the class {@link StringArrayFactory}.
+ * The unit test class for the class {@link LocalDateTimeFactory}.
  */
-public class StringArrayFactoryTest
+public class LocalDateTimeFactoryTest
 {
 
 	/**
-	 * Test for method {@link StringArrayFactory#create()}
+	 * Test for method {@link LocaleFactory#create()}
 	 */
-	@Test public void create()
+	@Test public void testCreate()
 	{
-		String[] expected;
-		String[] actual;
-		StringArrayFactory factory = new StringArrayFactory();
-		expected = new String[] { "foo", "bar", "john", "doe" };
+		LocalDateTime expected;
+		LocalDateTime actual;
+		LocalDateTimeFactory factory = new LocalDateTimeFactory();
+		expected = LocalDateTime.now();
 		actual = factory.create();
-		assertThat(actual, equalTo(expected));
+		assertEquals(actual.getYear(), expected.getYear());
+		assertEquals(actual.getMonth(), expected.getMonth());
+		assertEquals(actual.getDayOfMonth(), expected.getDayOfMonth());
+		assertEquals(actual.getHour(), expected.getHour());
+		assertEquals(actual.getMinute(), expected.getMinute());
+		assertEquals(actual.getSecond(), expected.getSecond());
 	}
+
 }