+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 ## Change log
 ----------------------
 
-Version 1.3-SNAPSHOT
+Version 2-SNAPSHOT
 -------------
 
+ADDED: 
 
+- added new factory class for the new java.util.time package LocalDate, LocalTime and LocalDateTime
+- added new test-dependency junit-jupiter (junit 5) in version 5.6.2
+
+CHANGED:
+
+- update jdk version to 11
+- update gradle to new version 6.5
+- update of dependency silly-collections to new version 8.2
+- remove of test-dependency testng 
+- extracted project properties to gradle.properties
 
 Version 1.2
 -------------